+++ tests/unit/experimental/pipeline/test_kg_builder.py
@@ -41,6 +41,7 @@ async def test_knowledge_graph_builder_init_with_text() -> None:
     assert kg_builder.entities == []
     assert kg_builder.relations == []
     assert kg_builder.potential_schema == []
+    assert "pdf_loader" not in kg_builder.pipeline
 
     text_input = "May thy knife chip and shatter."
 
@@ -72,6 +73,7 @@ async def test_knowledge_graph_builder_init_with_file_path() -> None:
     assert kg_builder.entities == []
     assert kg_builder.relations == []
     assert kg_builder.potential_schema == []
+    assert "pdf_loader" in kg_builder.pipeline
 
     file_path = "path/to/test.pdf"
 
@@ -241,3 +243,14 @@ def test_simple_kg_pipeline_on_error_invalid_value() -> None:
         )
 
     assert "Expected 'RAISE' or 'CONTINUE'" in str(exc_info.value)
+
+
+def test_simple_kg_pipeline_no_entity_resolution() -> None:
+    llm = MagicMock(spec=LLMInterface)
+    driver = MagicMock(spec=neo4j.Driver)
+
+    kg_builder = SimpleKGPipeline(
+        llm=llm, driver=driver, on_error="CONTINUE", perform_entity_resolution=False
+    )
+
+    assert "resolver" not in kg_builder.pipeline